Very basic (no operations!) 0th cut at sts.
Builds.  Doesn't crash when running.
Doesn't handle any ops at all; this is just the framework.

// fixup - merge with previous patch.

Marcus Watts <mwatts@redhat.com>
