+++ src/rgw/rgw_rest_s3.cc
@@ -252,7 +252,7 @@ done:
   if (!content_type)
     content_type = "binary/octet-stream";
 
-  end_header(s, this, content_type);
+  end_header(s, dump_access_control_f(), content_type);
 
   if (metadata_bl.length()) {
     STREAM_IO(s)->write(metadata_bl.c_str(), metadata_bl.length());
@@ -341,7 +341,7 @@ void RGWGetUsage_ObjStore_S3::send_response()
     set_req_state_err(s, op_ret);
   dump_errno(s);
 
-  end_header(s, this, "application/xml");
+  end_header(s, dump_access_control_f(), "application/xml");
   dump_start(s);
   if (op_ret < 0)
     return;
@@ -609,7 +609,7 @@ void RGWListBucket_ObjStore_S3::send_response()
 void RGWGetBucketLogging_ObjStore_S3::send_response()
 {
   dump_errno(s);
-  end_header(s, this, "application/xml");
+  end_header(s, dump_access_control_f(), "application/xml");
   dump_start(s);
 
   s->formatter->open_object_section_in_ns("BucketLoggingStatus",
@@ -621,7 +621,7 @@ void RGWGetBucketLogging_ObjStore_S3::send_response()
 void RGWGetBucketLocation_ObjStore_S3::send_response()
 {
   dump_errno(s);
-  end_header(s, this);
+  end_header(s, dump_access_control_f());
   dump_start(s);
 
   RGWZoneGroup zonegroup;
@@ -645,7 +645,7 @@ void RGWGetBucketLocation_ObjStore_S3::send_response()
 void RGWGetBucketVersioning_ObjStore_S3::send_response()
 {
   dump_errno(s);
-  end_header(s, this, "application/xml");
+  end_header(s, dump_access_control_f(), "application/xml");
   dump_start(s);
 
   s->formatter->open_object_section_in_ns("VersioningConfiguration",
@@ -831,7 +831,7 @@ void RGWStatBucket_ObjStore_S3::send_response()
   set_req_state_err(s, op_ret);
   dump_errno(s);
 
-  end_header(s, this);
+  end_header(s, dump_access_control_f());
   dump_start(s);
 }
 
@@ -993,7 +993,7 @@ void RGWDeleteBucket_ObjStore_S3::send_response()
 
   set_req_state_err(s, r);
   dump_errno(s);
-  end_header(s, this);
+  end_header(s, dump_access_control_f());
 
   if (s->system_request) {
     JSONFormatter f; /* use json formatter for system requests output */
@@ -1065,7 +1065,7 @@ void RGWPutObj_ObjStore_S3::send_response()
     dump_epoch_header(s, "Rgwx-Mtime", mtime);
   }
   dump_errno(s);
-  end_header(s, this);
+  end_header(s, dump_access_control_f());
 }
 
 /*
@@ -1856,7 +1856,7 @@ void RGWDeleteObj_ObjStore_S3::send_response()
   if (delete_marker) {
     dump_string_header(s, "x-amz-delete-marker", "true");
   }
-  end_header(s, this);
+  end_header(s, dump_access_control_f());
 }
 
 int RGWCopyObj_ObjStore_S3::init_dest_policy()
@@ -1978,7 +1978,7 @@ void RGWGetACLs_ObjStore_S3::send_response()
   if (op_ret)
     set_req_state_err(s, op_ret);
   dump_errno(s);
-  end_header(s, this, "application/xml");
+  end_header(s, dump_access_control_f(), "application/xml");
   dump_start(s);
   rgw_flush_formatter(s, s->formatter);
   STREAM_IO(s)->write(acls.c_str(), acls.size());
@@ -2024,7 +2024,7 @@ void RGWPutACLs_ObjStore_S3::send_response()
   if (op_ret)
     set_req_state_err(s, op_ret);
   dump_errno(s);
-  end_header(s, this, "application/xml");
+  end_header(s, dump_access_control_f(), "application/xml");
   dump_start(s);
 }
 
@@ -2336,7 +2336,7 @@ void RGWAbortMultipart_ObjStore_S3::send_response()
 
   set_req_state_err(s, r);
   dump_errno(s);
-  end_header(s, this);
+  end_header(s, dump_access_control_f());
 }
 
 void RGWListMultipart_ObjStore_S3::send_response()
@@ -2344,7 +2344,7 @@ void RGWListMultipart_ObjStore_S3::send_response()
   if (op_ret)
     set_req_state_err(s, op_ret);
   dump_errno(s);
-  end_header(s, this, "application/xml");
+  end_header(s, dump_access_control_f(), "application/xml");
 
   if (op_ret == 0) {
     dump_start(s);
@@ -2403,7 +2403,7 @@ void RGWListBucketMultiparts_ObjStore_S3::send_response()
     set_req_state_err(s, op_ret);
   dump_errno(s);
 
-  end_header(s, this, "application/xml");
+  end_header(s, dump_access_control_f(), "application/xml");
   dump_start(s);
   if (op_ret < 0)
     return;
@@ -2493,7 +2493,7 @@ void RGWDeleteMultiObj_ObjStore_S3::begin_response()
   }
 
   dump_start(s);
-  end_header(s, this, "application/xml");
+  end_header(s, dump_access_control_f(), "application/xml");
   s->formatter->open_object_section_in_ns("DeleteResult",
 					  "http://s3.amazonaws.com/doc/2006-03-01/");
