+++ src/rgw/rgw_rest_swift.cc
@@ -499,7 +499,7 @@ void RGWDeleteBucket_ObjStore_SWIFT::send_response()
 
   set_req_state_err(s, r);
   dump_errno(s);
-  end_header(s, this, NULL, 0);
+  end_header(s, dump_access_control_f(), NULL, 0);
   rgw_flush_formatter_and_reset(s, s->formatter);
 }
 
@@ -647,7 +647,7 @@ void RGWPutObj_ObjStore_SWIFT::send_response()
   dump_last_modified(s, mtime);
   set_req_state_err(s, op_ret);
   dump_errno(s);
-  end_header(s, this);
+  end_header(s, dump_access_control_f());
   rgw_flush_formatter_and_reset(s, s->formatter);
 }
 
@@ -705,7 +705,7 @@ void RGWPutMetadataAccount_ObjStore_SWIFT::send_response()
   }
   set_req_state_err(s, op_ret);
   dump_errno(s);
-  end_header(s, this);
+  end_header(s, dump_access_control_f());
   rgw_flush_formatter_and_reset(s, s->formatter);
 }
 
@@ -734,7 +734,7 @@ void RGWPutMetadataBucket_ObjStore_SWIFT::send_response()
   }
   set_req_state_err(s, op_ret);
   dump_errno(s);
-  end_header(s, this);
+  end_header(s, dump_access_control_f());
   rgw_flush_formatter_and_reset(s, s->formatter);
 }
 
@@ -767,7 +767,7 @@ void RGWPutMetadataObject_ObjStore_SWIFT::send_response()
     dump_content_length(s, 0);
   }
   dump_errno(s);
-  end_header(s, this);
+  end_header(s, dump_access_control_f());
   rgw_flush_formatter_and_reset(s, s->formatter);
 }
 
@@ -982,7 +982,7 @@ void RGWCopyObj_ObjStore_SWIFT::send_partial_response(off_t ofs)
       op_ret = STATUS_CREATED;
     set_req_state_err(s, op_ret);
     dump_errno(s);
-    end_header(s, this);
+    end_header(s, dump_access_control_f());
 
     /* Send progress information. Note that this diverge from the original swift
      * spec. We do this in order to keep connection alive.
